feat: :green_heart: Added GH runner
