+++ test/package.json
@@ -26,9 +26,9 @@
     "zombie:test:local": "SH_IMAGE=storage-hub:local pnpm zombienet test configs/simple.zndsl",
     "zombie:test:latest": "SH_IMAGE=docker.io/moonsonglabs/storage-hub:latest pnpm zombienet test configs/simple.zndsl",
     "test:full": "node --no-deprecation --import tsx --test ./suites/zombie/**.spec.ts",
-    "test:node": "node --no-deprecation --import tsx --test ./suites/solo-node/**/**.spec.ts",
-    "test:node:only": "node --no-deprecation --import tsx --test --test-only ./suites/solo-node/**/**.spec.ts",
-    "test:node:single": "node --no-deprecation --import tsx --test --test-name-pattern=$FILTER ./suites/solo-node/**/**.spec.ts"
+    "test:node": "node --no-deprecation --import tsx --test --test-reporter=@reporters/github --test-reporter-destination=stdout --test-reporter=spec --test-reporter-destination=stdout ./suites/solo-node/**/**.test.ts",
+    "test:node:only": "node --no-deprecation --import tsx --test --test-only ./suites/solo-node/**/**.test.ts",
+    "test:node:single": "node --no-deprecation --import tsx --test --test-name-pattern=$FILTER ./suites/solo-node/**/**.test.ts"
   },
   "keywords": [],
   "author": "",
@@ -39,14 +39,14 @@
     "@polkadot/api": "11.2.1",
     "@polkadot/api-augment": "11.2.1",
     "@polkadot/keyring": "12.6.2",
-    "@polkadot/types": "^11.2.1",
+    "@polkadot/types": "11.2.1",
     "@polkadot/util-crypto": "12.6.2",
     "@polkadot/wasm-crypto": "7.3.2",
+    "@reporters/github": "1.7.0",
     "@storagehub/api-augment": "link:../api-augment",
     "@unique-nft/sr25519": "0.0.2",
     "@zombienet/cli": "1.3.105",
     "@zombienet/utils": "0.0.25",
-    "expect": "29.7.0",
     "json-bigint": "^1.0.0",
     "lodash": "4.17.21",
     "testcontainers": "10.9.0",
