+++ test/suites/integration/test-1.test.ts
@@ -1,31 +1,32 @@
-import { expect } from "expect";
+import "core-js/stable";
+import { strictEqual } from "node:assert";
 import { randomInt } from "node:crypto";
 import { describe, it } from "node:test";
 
-describe("Suite 2", { only: true }, async () => {
-  it("Test 2.1", () => {
-    expect(true).toBe(true);
+describe("Suite 1", () => {
+  it("Test 1.1", () => {
+    strictEqual(true, true);
   });
 
-  it("Test 2.2", { only: true }, async () => {
-    const period = randomInt(0, 100) * 10;
+  it("Test 1.2", async () => {
+    const period = randomInt(0, 10) * 10;
     await new Promise((resolve) => setTimeout(resolve, period));
-    expect(true).toBe(true);
+    strictEqual(true, true);
   });
 
-  it("Test 2.3", async () => {
+  it("Test 1.3", async () => {
     const period = randomInt(0, 100) * 10;
     await new Promise((resolve) => setTimeout(resolve, period));
-    expect(true).toBe(true);
+    strictEqual(true, true);
   });
-  it("Test 2.4", async () => {
+  it("Test 1.4", async () => {
     const period = randomInt(0, 100) * 10;
     await new Promise((resolve) => setTimeout(resolve, period));
-    expect(true).toBe(true);
+    strictEqual(true, true);
   });
-  it("Test 2.5", async () => {
+  it("Test 1.5", async () => {
     const period = randomInt(0, 100) * 10;
     await new Promise((resolve) => setTimeout(resolve, period));
-    expect(true).toBe(true);
+    strictEqual(true, true);
   });
 });
