+++ test/suites/integration/test-4.test.ts
@@ -1,31 +1,31 @@
-import { expect } from "expect";
+import { strictEqual } from "node:assert";
 import { randomInt } from "node:crypto";
 import { describe, it } from "node:test";
 
-describe("Suite 3", () => {
-  it("Test 3.1", () => {
-    expect(true).toBe(true);
+describe("Suite 4", () => {
+  it("Test 4.1", () => {
+    strictEqual(true, true);
   });
 
-  it("Test 3.2", async () => {
+  it("Test 4.2", async () => {
     const period = randomInt(0, 100) * 10;
     await new Promise((resolve) => setTimeout(resolve, period));
-    expect(true).toBe(true);
+    strictEqual(true, true);
   });
 
-  it("Test 3.3", async () => {
+  it("Test 4.3", async () => {
     const period = randomInt(0, 100) * 10;
     await new Promise((resolve) => setTimeout(resolve, period));
-    expect(true).toBe(true);
+    strictEqual(true, true);
   });
-  it("Test 3.4", async () => {
+  it("Test 4.4", async () => {
     const period = randomInt(0, 100) * 10;
     await new Promise((resolve) => setTimeout(resolve, period));
-    expect(true).toBe(true);
+    strictEqual(true, true);
   });
-  it("Test 3.5", async () => {
+  it("Test 4.5", async () => {
     const period = randomInt(0, 100) * 10;
     await new Promise((resolve) => setTimeout(resolve, period));
-    expect(true).toBe(true);
+    strictEqual(true, true);
   });
 });
